code clean-up

+++ Seashore/source/document/XCFLayer.h
@@ -31,6 +31,6 @@
 				information).
 	@result		Returns instance upon success (or NULL otherwise).
 */
-- (id)initWithFile:(FILE *)file offset:(int)offset document:(id)doc sharedInfo:(SharedXCFInfo *)info;
+- (id)initWithFile:(FILE *)file offset:(long)offset document:(id)doc sharedInfo:(SharedXCFInfo *)info;
 
 @end
